+++ src/go/rpk/pkg/cli/security/acl/list.go
