+++ src/go/rpk/pkg/cli/security/user/delete.go
