+++ src/go/rpk/pkg/cli/security/user/list.go
